Apply rubocop recomendations

+++ decidim-surveys/db/migrate/20200610105927_remove_survey_columns.rb
@@ -2,17 +2,9 @@
 
 class RemoveSurveyColumns < ActiveRecord::Migration[5.2]
   def change
-    if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :title)
-      remove_column :decidim_surveys_surveys, :title
-    end
-    if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :description)
-      remove_column :decidim_surveys_surveys, :description
-    end
-    if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :tos)
-      remove_column :decidim_surveys_surveys, :tos
-    end
-    if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :published_at)
-      remove_column :decidim_surveys_surveys, :published_at
-    end
+    remove_column :decidim_surveys_surveys, :title if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :title)
+    remove_column :decidim_surveys_surveys, :description if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :description)
+    remove_column :decidim_surveys_surveys, :tos if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :tos)
+    remove_column :decidim_surveys_surveys, :published_at if ActiveRecord::Base.connection.column_exists?(:decidim_surveys_surveys, :published_at)
   end
 end
